signaling_layer/requirements.txt
fastapi
uvicorn[standard]
redis
grpcio
grpcio-tools
protobuf
websockets
python-multipart
ujson
